refactor: 코드 컨벤션 준수 및 패키지 분리
racinggame.domain.car 내에 generatenumber 패키지 분리
ALL 코드 컨벤션 수정
