+++ src/main/java/racinggame/domain/car/generatenumber/GenerateNumberService.java
@@ -1,4 +1,4 @@
-package racinggame.domain.car;
+package racinggame.domain.car.generatenumber;
 
 public interface GenerateNumberService {
     public int generateNumber(int bound);
