+++ src/main/java/racinggame/domain/car/generatenumber/GenerateRandomNumber.java
@@ -1,9 +1,9 @@
-package racinggame.domain.car;
+package racinggame.domain.car.generatenumber;
 
 import java.util.Random;
 
 public class GenerateRandomNumber implements GenerateNumberService {
-    public int generateNumber(int bound){
+    public int generateNumber(int bound) {
         return new Random(System.currentTimeMillis()).nextInt(bound);
     }
 }
