+++ src/test/java/racinggame/domain/GenerateTestNumber.java
@@ -1,9 +1,9 @@
 package racinggame.domain;
 
-import racinggame.domain.car.GenerateNumberService;
+import racinggame.domain.car.generatenumber.GenerateNumberService;
 
 public class GenerateTestNumber implements GenerateNumberService {
-    public int generateNumber(int bound){
+    public int generateNumber(int bound) {
         return bound - 1;
     }
 }
